Update the data types for play asset stats.

+++ Grunt/Grunt/Models/HaloInfinite/PlayAssetStats.cs
@@ -7,16 +7,50 @@
 
 namespace OpenSpartan.Grunt.Models.HaloInfinite
 {
+    /// <summary>
+    /// Class containing information about asset statistics.
+    /// </summary>
     [IsAutomaticallySerializable]
     public class PlayAssetStats
     {
-        public int? PlaysRecent { get; set; }
-        public int? PlaysAllTime { get; set; }
-        public int? Favorites { get; set; }
-        public int? Likes { get; set; }
-        public int? Bookmarks { get; set; }
-        public int? ParentAssetCount { get; set; }
+        /// <summary>
+        /// Gets or sets the number of recent plays for an asset.
+        /// </summary>
+        public long? PlaysRecent { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of all-time plays for an asset.
+        /// </summary>
+        public long? PlaysAllTime { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of times an asset has been favorited.
+        /// </summary>
+        public long? Favorites { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of times an asset has been liked.
+        /// </summary>
+        public long? Likes { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of times an asset has been bookmarked.
+        /// </summary>
+        public long? Bookmarks { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of assets that are parented from the current asset.
+        /// </summary>
+        public long? ParentAssetCount { get; set; }
+
+        /// <summary>
+        /// Gets or sets the average rating for an asset.
+        /// </summary>
         public float? AverageRating { get; set; }
-        public int? NumberOfRatings { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of ratings for an asset.
+        /// </summary>
+        public long? NumberOfRatings { get; set; }
     }
 }
